Make self-repair static and remove recipe previews.
