+++ recipe-server/client/actions/tests/utils.js
@@ -1,6 +1,54 @@
 import uuid from 'node-uuid';
 
-import { HeartbeatEmitter } from '../../selfrepair/normandy_driver.js';
+/**
+ * EventEmitter for Heartbeat events. Unlike normal event emitters,
+ * Heartbeat events only occur once, and we guarantee that if a handler
+ * is registered _after_ an event occurs, it will still be executed.
+ */
+export class HeartbeatEmitter {
+  static EVENTS = [
+    'NotificationOffered',
+    'NotificationClosed',
+    'LearnMore',
+    'Voted',
+    'TelemetrySent',
+    'Engaged',
+  ];
+
+  constructor() {
+    this.callbacks = {};
+    this.eventData = {};
+    for (const event of HeartbeatEmitter.EVENTS) {
+      this.callbacks[event] = [];
+      this.eventData[event] = null;
+    }
+  }
+
+  on(eventName, callback) {
+    if (HeartbeatEmitter.EVENTS.indexOf(eventName) === -1) {
+      throw new Error(`${eventName} is an invalid Heartbeat event type.`);
+    }
+
+    this.callbacks[eventName].push(callback);
+
+    // Call the callback if the event already happened.
+    const data = this.eventData[eventName];
+    if (data !== null) {
+      callback(data);
+    }
+  }
+
+  emit(eventName, data) {
+    if (this.eventData[eventName] !== null) {
+      throw new Error(`Cannot emit a Heartbeat event more than once: ${eventName}`);
+    }
+
+    this.eventData[eventName] = data;
+    for (const callback of this.callbacks[eventName]) {
+      callback(data);
+    }
+  }
+}
 
 export class MockStorage {
   constructor() {
