--- recipe-server/client/selfrepair/tests/.eslintrc
@@ -1,3 +1,0 @@
-env:
-  jasmine: true
-  browser: true
