--- recipe-server/client/selfrepair/tests/index.js
@@ -1,2 +1,0 @@
-const testsContext = require.context('./', true, /test_[^//]*\.js$/);
-testsContext.keys().forEach(testsContext);
