+++ recipe-server/normandy/selfrepair/tests/test_views.py
@@ -1,5 +1,3 @@
-from urllib.parse import urljoin
-
 from django.core.urlresolvers import reverse
 from django.db import connection
 from django.test.utils import CaptureQueriesContext
@@ -7,7 +5,7 @@ from django.test.utils import CaptureQueriesContext
 import pytest
 
 
-class TestSelfRepair:
+class TestSelfRepair(object):
     url = '/en-US/repair'
 
     def test_url_is_right(self):
@@ -32,19 +30,3 @@ class TestSelfRepair:
         res = client.get(self.url)
         assert res.status_code == 200
         assert res.client.cookies == {}
-
-    def test_classify_url_is_right_no_app_server(self, client, settings):
-        settings.APP_SERVER_URL = None
-        classify_client_url = reverse('recipes:classify-client')
-        res = client.get(self.url)
-        assert res.status_code == 200
-        expected = 'data-classify-url="{}"'.format(classify_client_url)
-        assert expected in res.content.decode()
-
-    def test_classify_url_is_right_with_app_server(self, client, settings):
-        settings.APP_SERVER_URL = 'https://testserver-cdn/'
-        classify_client_url = urljoin(settings.APP_SERVER_URL, reverse('recipes:classify-client'))
-        res = client.get(self.url)
-        assert res.status_code == 200
-        expected = 'data-classify-url="{}"'.format(classify_client_url)
-        assert expected in res.content.decode()
